+++ integration-service-module/src/main/java/cz/mamrakm/integrationservice/AccountProcessor.java
@@ -1,19 +1,18 @@
 package cz.mamrakm.integrationservice;
 
-import lombok.Builder;
-import lombok.ToString;
-import lombok.extern.slf4j.Slf4j;
+import cz.mamrakm.integrationservice.dtos.AccountDTO;
 import org.apache.camel.Exchange;
 import org.apache.camel.Processor;
 import org.mapstruct.factory.Mappers;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
-@Slf4j
-@Builder
-@ToString
 public class AccountProcessor implements Processor {
-    @Override
-    public void process(Exchange exchange) {
-        CommonModelMapper commonModelMapper= Mappers.getMapper(CommonModelMapper.class);
-        log.info(">>> body: {}", exchange.getMessage().getBody().toString());
-    }
+  @Override
+  public void process(Exchange exchange) {
+    Logger logger = LoggerFactory.getLogger(AccountProcessor.class);
+    CommonModelMapper commonModelMapper = Mappers.getMapper(CommonModelMapper.class);
+    AccountDTO integrationApiModel = AccountDTO.class.cast(exchange.getMessage().getBody());
+    logger.info(">>> body: {}", integrationApiModel.getFirstname());
+  }
 }