+++ integration-service-module/src/main/java/cz/mamrakm/integrationservice/CommonModelMapper.java
@@ -3,11 +3,14 @@ package cz.mamrakm.integrationservice;
 import cz.mamrakm.integrationservice.dtos.AccountDTO;
 import cz.mamrakm.integrationservice.models.CommonModel;
 import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
 
 @Mapper
 public interface CommonModelMapper {
 
-    @Mapping(target = "CommonModel.apiModel")
-    CommonModel accountToCommon(AccountDTO accountDTO);
+  default CommonModel accountToCommon(AccountDTO accountDTO) {
+    CommonModel commonModel = new CommonModel();
+    commonModel.setOperation("POST");
+    commonModel.setApiModel(accountDTO);
+    return commonModel;
+  }
 }