+++ integration-service-module/src/main/java/cz/mamrakm/integrationservice/IntegrationService.java
@@ -1,10 +1,6 @@
 package cz.mamrakm.integrationservice;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
-@RequiredArgsConstructor
-@Service
 public class IntegrationService {
 
 }