+++ integration-service-module/src/main/java/cz/mamrakm/integrationservice/IntegrationServiceApplication.java
@@ -1,20 +1,22 @@
 package cz.mamrakm.integrationservice;
 
-import lombok.extern.slf4j.Slf4j;
 import org.apache.camel.Processor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ImportResource;
 
-@Slf4j
 @ImportResource("classpath:integration-bean.xml")
 @SpringBootApplication
 public class IntegrationServiceApplication {
 
   public static void main(String[] args) {
+    Logger logger = LoggerFactory.getLogger(IntegrationServiceApplication.class);
+
     SpringApplication.run(IntegrationServiceApplication.class, args);
-    log.trace("IntegrationServiceApplication started");
+    logger.info("IntegrationServiceApplication started");
   }
 
   @Bean