+++ integration-service-module/src/main/java/cz/mamrakm/integrationservice/JSONMapper.java
@@ -1,3 +1,0 @@
-package cz.mamrakm.integrationservice;
-
-public class JSONMapper {}