+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2] - 2023-05-23
+
+### Removed
+
+Lombok
+
+It doesn't work well with
+Mapstruct [Stackoverflow issue](https://stackoverflow.com/questions/34358689/maven-build-cannot-find-symbol-when-accessing-project-lombok-annotated-methods)
+also some (like me) consider using Lombok a bad practice.
+
 ## [0.0.1] - 2023-05-15
+
 ### Added
+
 project initialization